Ensure ARTEMIS config file is not clobbered on pod restart (#232)
Previously when the backend pod restarted it would overwrite the config file in persistent storage with the one from the configmap (i.e., the default). The `-u` option to `cp` doesn't seem to be enough to prevent an overwrite (presumably the modification time of the file from the configmap cannot be trusted). Using both `-u` and the `-n` (no-clobber) option seems to work and prevent the persisted config file from being overwritten.