Fix file for easier use  (#609) [ci skip]
* Update bgpstream_retrieve_prefix_records.py

change from BGPstream v1 to v2.

* Create install_pybgpstreamv2_for_ubuntu

* Rename install_pybgpstreamv2_for_ubuntu to install_pybgpstreamv2_for_ubuntu.sh

* Update bgpstream_retrieve_prefix_records.py

* Update bgpstream_retrieve_prefix_records.py

* Rename json_file to json_dump; use of pre-commit

* Delete install_pybgpstreamv2_for_ubuntu.sh

The file is removed because it seems to cause errors in the automatic validation process.

* Update logging.yaml

If the #s were removed from the handlers, they were at the wrong distance from the other handlers. This led to an error in the yaml interpreter. 
By removing a space, the commands are in the correct position after removing the #.